backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
google-adk
google-genai
neo4j
neo4j-graphrag[google,experimental,fuzzy-matching]
rdflib
youtube-transcript-api
beautifulsoup4
requests
pandas
pydantic
